fix all coercion
